requirements.txt
cachetools==4.2.4
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.4
dataclasses==0.8
dnspython==2.2.1
Flask==2.0.3
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
google-api-core==2.8.0
google-api-python-client==2.48.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.1
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
oauthlib==3.2.0
pkg-resources==0.0.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==4.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
zipp==3.6.0
